refactor(commonpb): colocate metric json encoding
