+++ lint/style/allowlist.txt
@@ -193,7 +193,6 @@ testpkg:bindings/go/dataplane_ut/dataplane_ut_test.go:dataplaneut:1  # legacy: m
 testpkg:bindings/go/filter/types_test.go:filter:1  # legacy: migrate to the external test package
 testpkg:common/commonpb/v1/ipaddr_test.go:commonpb:1  # legacy: migrate to the external test package
 testpkg:common/commonpb/v1/iprange_test.go:commonpb:1  # legacy: migrate to the external test package
-testpkg:common/commonpb/v1/json_test.go:commonpb:1  # legacy: migrate to the external test package
 testpkg:common/commonpb/v1/macaddr_test.go:commonpb:1  # legacy: migrate to the external test package
 testpkg:common/commonpb/v1/metrics_test.go:commonpb:1  # legacy: migrate to the external test package
 testpkg:common/go/bitset/bitset_test.go:bitset:1  # legacy: migrate to the external test package
